+++ src/lib/Archive/Types/ValueToJSON.cpp
@@ -1,6 +1,6 @@
 #include "ValueToJSON.hpp"
 
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 #include <Error/Exception.hpp>
 #include <Specification/Core/CoreModule.hpp>
