+++ src/lib/Messaging/Method.cpp
@@ -1,6 +1,6 @@
 #include "Method.hpp"
 
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 #include <Error/Exception.hpp>
