+++ src/lib/Messaging/TokenStream.hpp
@@ -2,7 +2,7 @@
 
 #include "Token.hpp"
 
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 
 
 namespace sedmgr {
