+++ src/lib/MockDevice/MockDevice.cpp
@@ -2,7 +2,7 @@
 
 #include "Preconfig.hpp"
 
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 #include <Error/Exception.hpp>
 #include <Messaging/TokenStream.hpp>
 #include <Messaging/Value.hpp>
