+++ src/lib/TrustedPeripheral/Discovery.cpp
@@ -1,6 +1,6 @@
 #include "Discovery.hpp"
 
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 
 #include <cereal/archives/portable_binary.hpp>
