+++ src/lib/TrustedPeripheral/MethodUtils.cpp
@@ -2,7 +2,7 @@
 
 #include "Logging.hpp"
 
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 #include <Error/Exception.hpp>
 #include <Messaging/TokenStream.hpp>
