+++ src/lib/TrustedPeripheral/Session.cpp
@@ -3,7 +3,7 @@
 #include "Logging.hpp"
 #include <async++/join.hpp>
 
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 #include <Error/Exception.hpp>
 #include <Messaging/ComPacket.hpp>
 #include <Messaging/TokenStream.hpp>
