+++ src/lib/TrustedPeripheral/TrustedPeripheral.cpp
@@ -4,7 +4,7 @@
 #include <async++/join.hpp>
 #include <async++/sleep.hpp>
 
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 #include <Error/Exception.hpp>
 #include <Messaging/ComPacket.hpp>
 #include <Messaging/SetupPackets.hpp>
