+++ test/Messaging/TestTokenStream.cpp
@@ -1,4 +1,4 @@
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 #include <Messaging/TokenStream.hpp>
 
 #include <catch2/catch_test_macros.hpp>
