+++ test/TrustedPeripheral/TestDiscovery.cpp
@@ -1,4 +1,4 @@
-#include <Archive/Conversion.hpp>
+#include <Archive/Serialization.hpp>
 #include <TrustedPeripheral/Discovery.hpp>
 
 #include <catch2/catch_test_macros.hpp>
